requirements.txt
# Modern Data Warehouse - Python Dependencies
# ==========================================

# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Database connectivity
pyodbc>=4.0.35
sqlalchemy>=1.4.0

# Configuration management
pyyaml>=6.0

# Logging and monitoring
structlog>=22.0.0

# Data validation and quality
great-expectations>=0.15.0
cerberus>=1.3.4

# Scheduling and automation
schedule>=1.2.0
croniter>=1.3.0

# Visualization and reporting (optional)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.990

# Data profiling
pandas-profiling>=3.6.0

# Email notifications (optional)
smtplib2>=0.2.1

# Progress bars
tqdm>=4.64.0

# Date/time utilities
python-dateutil>=2.8.0

# Environment management
python-dotenv>=0.20.0

# Performance monitoring
psutil>=5.9.0

# File handling
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# Web framework for monitoring dashboard (optional)
flask>=2.2.0
dash>=2.6.0